Fix javadoc error caused by mismatched package paths
