+++ evaluator.measure-hapi/pom.xml
@@ -4,7 +4,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>org.opencds.cqf.cql</groupId>
-    <artifactId>evaluator.measure.hapi</artifactId>
+    <artifactId>evaluator.measure-hapi</artifactId>
     <version>1.2.1-SNAPSHOT</version>
     <packaging>jar</packaging>
